add: one service to handle API logic.
